+++ etc/scripts/gen_requirements_dev.py
@@ -36,7 +36,7 @@ def gen_dev_requirements():
         type=pathlib.Path,
         required=True,
         metavar="DIR",
-        help='Path to the "site-packages" directory where wheels are installed such as lib/python3.6/site-packages',
+        help='Path to the "site-packages" directory where wheels are installed such as lib/python3.7/site-packages',
     )
     parser.add_argument(
         "-d",