+++ src/attributecode/__init__.py
@@ -20,9 +20,9 @@ import os
 
 import saneyaml
 
-__version__ = '7.2.1'
+__version__ = '8.0.0'
 
-__about_spec_version__ = '3.2.3'
+__about_spec_version__ = '3.3.0'
 
 __copyright__ = """
 Copyright (c) nexB Inc. All rights reserved. http://dejacode.org